+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ## [Unreleased]
 
+### Added
+- Mouse hovering variable placeholders in Body Editor to show a tooltip for its value (if exists)
+
 ### Removed
 - Text fields and response body viewer now do not trim content over 4 MB (but other limits still apply)